\name{SPIDDOR-package}
\Rdversion{1.0}
\alias{SPIDDOR-package}
\alias{SPIDDOR}
\docType{package}
\title{
\packageTitle{SPIDDOR}
}
\description{
\packageDescription{SPIDDOR}
Modeling, simulation and analysis of Boolean networks applied to Systems Pharmacology.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SPIDDOR}
\packageIndices{SPIDDOR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SPIDDOR}
Autors: Itziar Irurzun-Arana, Jose David Gomez-Mantilla, Iñaki Troconiz.
Maintainer: Itziar Irurzun <itzirurzun@alumni.unav.es>\packageMaintainer{SPIDDOR}
}
\references{
Kauffman, S. A. 1969. "Metabolic Stability and Epigenesis in Randomly Constructed Genetic Nets." Journal of Theoretical Biology 22.
}

\keyword{ package
Boolean network
attractor
Systems Pharmacology
Simulation}

\seealso{
~~ Optional links to other man pages, e.g. ~~

}
\examples{
%% ~~ simple examples of the most important functions ~~
}
