\name{SPIDDOR-package}
\Rdversion{1.0}
\alias{SPIDDOR-package}
\alias{SPIDDOR}
\docType{package}
\title{
\packageTitle{SPIDDOR}
}
\description{
\packageDescription{SPIDDOR}
}
\details{
SPIDDOR is an R package which consist on a set of tools to perform Boolean modeling in the context of development therapies for complex diseases. SPIDDOR allows users to simulate synchronous and asynchronous Boolean networks and analyze the results in terms of the average dynamic evolution of the nodes or in terms of attractors.

From a methodological point of view the Boolean analysis performed by SPIDDOR involves certain novelties. Common Boolean modeling approaches only define direct activation-inhibition relationships between the components of the network. In our models, we incorporate the modulation interactions, which are used to modulate the intensity of the activations or inhibitions produced by the regulator nodes. The package also allows users to specify the activity level of their nodes as a percentage ON in order to perform mutational studies or evaluate the inputs of the network with background noises. Additionally, SPIDDOR incorporates new visualization techniques to evaluate the attractors of the system and the effects of perturbations.

\packageDESCRIPTION{SPIDDOR}
\packageIndices{SPIDDOR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Authors: \packageAuthor{SPIDDOR}

Maintainer: \packageMaintainer{SPIDDOR}
}
\references{
Kauffman, S. A. 1969. "Metabolic Stability and Epigenesis in Randomly Constructed Genetic Nets." Journal of Theoretical Biology 22.
}

\keyword{ package
Boolean network
attractor
Systems Pharmacology
Simulation}

\examples{
%% ~~ simple examples of the most important functions ~~
}
